update 'validation' package
Unit testing of validation code.

Factored out 'messages.go' to hold the JSON struct types for clearer unit testing.

Installing packages without having to list them in the travis file as well.
