+++ .travis.yml
@@ -1,13 +1,16 @@
 language: go
 
 go:
-  - 1.10.x
-  - 1.11.x
-  - 1.12.x
+  - "1.10"
+  - "1.11"
+  - "1.12"
+  - "1.13"
+  - "1.14"
+  - "1.15"
+  - "1.16"
 
-before_install:
-  - go get -v github.com/gorilla/mux
-  - go get -v gopkg.in/go-playground/validator.v9
+install:
+  - make install
 
 script:
   - make test
