+++ main.go
@@ -2,43 +2,17 @@ package main
 
 import (
 	"encoding/json"
-	"github.com/gorilla/mux"
-	"gopkg.in/go-playground/validator.v9"
 	"io/ioutil"
 	"log"
 	"net/http"
 	"path"
 	"strconv"
 	"time"
+
+	"github.com/gorilla/mux"
 )
 
-var validate *validator.Validate
-
-func ValidateMute(fl validator.FieldLevel) bool {
-	switch fl.Field().String() {
-	case
-		"mute",
-		"unmute":
-		return true
-	}
-	return false
-}
-
-func ValidateVolume(fl validator.FieldLevel) bool {
-	switch fl.Field().String() {
-	case
-		"up",
-		"down":
-		return true
-	}
-	return false
-}
-
 func main() {
-	validate = validator.New()
-	validate.RegisterValidation("mute", ValidateMute)
-	validate.RegisterValidation("volume", ValidateVolume)
-
 	var port = 8080
 	log.Print("Starting - http://localhost:" + strconv.Itoa(port) + "/")
 
@@ -88,18 +62,6 @@ func pingHandler(w http.ResponseWriter, r *http.Request) {
 	log.Print("ping")
 	w.Header().Add("Content-Type", "application/json")
 	w.Write([]byte(`true`))
-}
-
-type ErrorResponse struct {
-	Message string `json:"message"`
-}
-
-type MuteRequest struct {
-	Mute string `json:"mute" validate:"mute"`
-}
-
-type VolumeRequest struct {
-	Volume string `json:"volume" validate:"volume"`
 }
 
 func setMuteHandler(w http.ResponseWriter, r *http.Request) {
